// src/app/Shared/models/life/setup/organization/appraisal/criteria.ts
export class Criteria {
  "authBy": string;
  "authOn": Date;
  "companyCode": string;
  "criteria1": string;
  "criteria2": string;
  "criteria3": string;
  "criteria4": string;
  "criteria5": string;
  "e1Charge": number;
  "e1time": number;
  "e2Charge": number;
  "errorGCharge": number;
  "errorICharge": number;
  "errorMCharge": number;
  "versionNo": string;
}